feat(docs): Add 004-action-pattern ADR and fix action header
