+++ docs/adr/README.md
@@ -15,3 +15,4 @@ Minimum required sections: **Context**, **Decision**, **Consequences**
 | [001](/adr/001-ecs-with-flecs.md) | ECS with Flecs | Accepted |
 | [002](/adr/002-stats-and-modifiers.md) | Stats and Modifiers | Accepted |
 | [003](/adr/003-finance-system.md) | Finance System | Proposed |
+| [004](/adr/004-action-pattern.md) | UI Action Pattern | Accepted |
